requirements.txt
dbt-duckdb==1.7.1
duckdb==0.10.0
python-dotenv==1.0.1
requests==2.32.5